Finally exam

+++ src/service/staff.service.js
@@ -1,46 +1,39 @@
 import newStaff from "../models/staff.model.js";
-import CustomError from "../utils/error.js"
+import CustomError from "../utils/error.js";
 import sha256 from "sha256";
-import JWT from "../utils/jwt.js"
+import JWT from "../utils/jwt.js";
 
-export class StaffService{
+export class StaffService {
 
-    constructor(){}
+    constructor() {}
 
-    static async CreateStaff(body,dataToken){
-        try{
-            let odam = await newStaff.findOne({username:body.username})
-            if(odam) throw new CustomError(402,"Bunday User bor")  
-           
+    static async CreateStaff(body, dataToken) {
+        try {
+            let odam = await newStaff.findOne({ username: body.username });
+            if (odam) throw new CustomError(402, "Bunday User bor");
+
             body.password = sha256(body.password);
-               
-                delete body.repeat_password
-                
-            let top = await newStaff.create(body)
-    
+            delete body.repeat_password;
+
+            let top = await newStaff.create(body);
             dataToken._id = top._id;
-    
+
             return {
                 accessToken: JWT.sign(dataToken),
                 refreshToken: JWT.signRef(dataToken)
-                
             };
         } catch (error) {
-            throw new CustomError(error.status || 500,error.message || "Bunday foydalanuvchi bor");
+            throw new CustomError(error.status || 500, error.message || "Foydalanuvchi yaratishda xatolik");
         }
     }
-
-
-    
 
     static async loginUser(body, dataToken) {
         try {
             let { username, password } = body;
-
             let foundUser = await newStaff.findOne({ username, password: sha256(password) });
 
             if (!foundUser) {
-                throw new CustomError(404,"Bunday foydalanuvchi topilmadi");
+                throw new CustomError(404, "Bunday foydalanuvchi topilmadi");
             }
 
             dataToken._id = foundUser._id;
@@ -51,77 +44,73 @@ export class StaffService{
                 foundUser
             };
         } catch (error) {
-            throw new CustomError(error.status || 500, error.message);
+            throw new CustomError(error.status || 500, error.message || "Login xatoligi");
         }
     }
 
     static async userAll() {
         try {
-            let users = await newStaff.find();
+            let users = await newStaff.find({ role: "Staff" });
 
             return {
                 message: "Success",
                 users
             };
-
         } catch (error) {
-            throw new CustomError(error.status || 500,"Foydalanuvchilarni olishda xatolik");
+            throw new CustomError(error.status || 500, "Foydalanuvchilarni olishda xatolik");
         }
     }
-
 
     static async userOne(body) {
         try {
-    
-            let users = await newStaff.findById(body);
-    
+            let users = await newStaff.findOne({ _id: body.id, role: "Staff" });
+
             if (!users) {
-                throw new CustomError(404 ,"Foydalanuvchi topilmadi");
+                throw new CustomError(404, "Foydalanuvchi topilmadi");
             }
-    
+
             return {
                 message: "Success",
                 users
             };
-    
         } catch (error) {
-            throw new CustomError(error.status || 500,"Foydalanuvchilarni olishda xatolik");
+            throw new CustomError(error.status || 500, "Foydalanuvchini olishda xatolik");
         }
     }
-    
+
     static async userDelete(body) {
         try {
-            let users = await newStaff.deleteOne({_id:body.id});
+            let users = await newStaff.deleteOne({ _id: body.id, role: "Staff" });
 
             return {
                 message: "User delete",
-                succase:"Success"
+                success: "Success"
             };
-
         } catch (error) {
-            console.log(error);
-            throw new CustomError(error.status || 500,"Foydalanuvchilarni olishda xatolik");
+            throw new CustomError(error.status || 500, "Foydalanuvchini o‘chirishda xatolik");
         }
     }
-    
+
     static async userUpdate(id, data) {
         try {
-           let user = await newStaff.findById(id)         
-           if(!user){
-             throw new CustomError(error.status || 500,"User not found")
-           };
-           if (data.username) user.username = data.username
-           if (data.password) user.password = data.password
-           if (data.birthDate) user.birthDate = data.birthDate
-           user.save()
+            let user = await newStaff.findOne({ _id: id, role: "Staff" });
+
+            if (!user) {
+                throw new CustomError(404, "Foydalanuvchi topilmadi");
+            }
+
+            if (data.username) user.username = data.username;
+            if (data.password) user.password = sha256(data.password);
+            if (data.birthDate) user.birthDate = data.birthDate;
+
+            await user.save();
+
             return {
                 message: "Foydalanuvchi muvaffaqiyatli yangilandi",
                 success: true
             };
         } catch (error) {
-            throw new CustomError(error.status || 500,"User not found");
-
+            throw new CustomError(error.status || 500, error.message || "Yangilashda xatolik");
         }
     }
-} 
-
+}
